% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkDoublets.R
\name{MarkDoublets}
\alias{MarkDoublets}
\title{Detect doublet cells using Scrublet}
\usage{
MarkDoublets(seu, PN = NULL)
}
\arguments{
\item{seu}{A Seurat object containing RNA count data}

\item{PN}{Expected doublet rate. If NULL, automatically calculated using calculate_pN()}
}
\value{
A seurat object whose metadata contain doublet scores and predictions for each barcode.
}
\description{
Performs doublet detection by wrapping the Python Scrublet package via reticulate.
}
