% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_process.R
\name{pcs_determine}
\alias{pcs_determine}
\title{Determine Significant Principal Components}
\usage{
pcs_determine(seu, embedding = "pca")
}
\arguments{
\item{seu}{Seurat object}

\item{embedding}{Name of reduction embedding to use (default: "pca")}
}
\value{
Integer indicating the number of PCs to retain
}
\description{
Calculates the number of PCs needed to explain 90% of cumulative variance.
}
\examples{
# Using default PCA embedding:
pcs <- pcs_determine(seu)

# Using a custom embedding:
pcs <- pcs_determine(seu, embedding = "custom_pca")

}
