% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_qc.R
\name{sample_qc}
\alias{sample_qc}
\title{Sample Quality Control}
\usage{
sample_qc(sc)
}
\arguments{
\item{sc}{A Seurat object containing single-cell RNA sequencing data}
}
\value{
Modified Seurat object with added QC metrics in meta.data:
  - percent_mito_by_sample (if MT genes present)
  - percent_ribo_by_sample (if RP genes present)
  - percent_hb_by_sample (if HB genes present)
  - percent_hsp_by_sample (if HSP genes present)
  - percent_dissociation_by_sample (if dissociation genes present)
}
\description{
Calculates various quality control metrics for each sample including
mitochondrial, ribosomal, hemoglobin, heat shock protein, and dissociation
gene percentages. Adds these metrics as new columns to the meta.data slot
of the input Seurat object.
}
\examples{
seurat_obj <- sample_qc(seurat_obj)
}
